refactor: 로그 화면의 툴바, 복사 로직을 액티비티로 이동
